Move `test_repeated_custom_type()` test to a correct location
