+++ tests/test_edge_cases.py
@@ -21,10 +21,18 @@ from typing import (
 
 import pytest
 from dirty_equals import HasRepr, IsStr
-from pydantic_core import ErrorDetails, InitErrorDetails, PydanticSerializationError, PydanticUndefined, core_schema
-from typing_extensions import TypeAliasType, TypedDict, get_args
+from pydantic_core import (
+    CoreSchema,
+    ErrorDetails,
+    InitErrorDetails,
+    PydanticSerializationError,
+    PydanticUndefined,
+    core_schema,
+)
+from typing_extensions import Self, TypeAliasType, TypedDict, get_args
 
 from pydantic import (
+    AfterValidator,
     BaseModel,
     ConfigDict,
     GetCoreSchemaHandler,
@@ -3008,6 +3016,38 @@ def test_get_pydantic_core_schema_on_referenceable_type() -> None:
         t: 'Test'
 
     assert counter == 1
+
+
+def test_repeated_custom_type() -> None:
+    class Numeric(BaseModel):
+        value: float
+
+        @classmethod
+        def __get_pydantic_core_schema__(cls, source_type: Any, handler: GetCoreSchemaHandler) -> CoreSchema:
+            return core_schema.no_info_before_validator_function(cls._validate, handler(source_type))
+
+        @classmethod
+        def _validate(cls, v: Any) -> Union[dict[str, Any], Self]:
+            if isinstance(v, (str, float, int)):
+                return cls(value=v)
+            if isinstance(v, Numeric):
+                return v
+            if isinstance(v, dict):
+                return v
+            raise ValueError(f'Invalid value for {cls}: {v}')
+
+    def is_positive(value: Numeric):
+        assert value.value > 0.0, 'Must be positive'
+
+    class OuterModel(BaseModel):
+        x: Numeric
+        y: Numeric
+        z: Annotated[Numeric, AfterValidator(is_positive)]
+
+    assert OuterModel(x=2, y=-1, z=1)
+
+    with pytest.raises(ValidationError):
+        OuterModel(x=2, y=-1, z=-1)
 
 
 def test_validator_and_serializer_not_reused_during_rebuild() -> None:
